Don't run ELS on parser
There is already elasticsearch deployed as a part of parser job. No
need to run another elasticsearch on the same VM. Moreover, ELS is
configured to consume 46% of the VM memory, which in case of 2 ELSes
means that the VM runs out of memory, starts swapping heavily and
becomes unresponsive.